refactor: move sgidloginpage into loginpage
