+++ src/routing/RouteSelector.jsx
@@ -14,7 +14,6 @@ import EditNavBar from "layouts/EditNavBar"
 import EditPage from "layouts/EditPage"
 import { Folders } from "layouts/Folders"
 import { LoginPage } from "layouts/Login"
-import { LoginPageSgid } from "layouts/Login/LoginPageSgid"
 import { Media } from "layouts/Media"
 import NotFoundPage from "layouts/NotFoundPage"
 import { ResourceCategory } from "layouts/ResourceCategory"
@@ -46,12 +45,10 @@ export const RouteSelector = () => (
       </Banner>
     )}
     <Switch>
-      <RedirectIfLoggedInRoute exact path="/" unauthedComponent={LoginPage} />
-
       <RedirectIfLoggedInRoute
         exact
-        path="/ogp-login"
-        unauthedComponent={LoginPageSgid}
+        path={["/", "/ogp-login"]}
+        unauthedComponent={LoginPage}
       />
 
       <ProtectedRouteWithProps
